feat: add option for populating ecommerce categories
